numactl - fix numastat sysfs scanning
Against:  numactl-1.0.3-rc1

On newer kernels, readdir may enumerate the 'node(\d+) subdirs in
the opposite order from older kernels--e.g., node{0,1,2,...} as
opposed to node{N,N-1,N-2,...}.  Not sure when this started
happening, but RHEL5 [~2.6.18+] behaves as an "older kernel",
while 2.6.25-rc2/3 behave in the "new way".

Modify numastat to detect when node0 is reported first, and swap
the order so that we continue to report numastats with node0
first--i.e., as leftmost node.

Note:   this still assumes that the nodes are listed in numeric
node id order--either increasing or decreasing.  A more ambitious
rewrite to save the per node stats in arrays indexed by node id
would be more robust to violations of this assumption.  Maybe later?

Signed-off-by:  Lee Schermerhorn <lee.schermerhorn@hp.com>

 numastat |   22 ++++++++++++++++++++--
 1 file changed, 20 insertions(+), 2 deletions(-)