numactl/libnuma: version 1 source code compatibility
Andi suggested we supply some compatibility wrappers with a compat
define for old programs.

This patch provides numacompat1.h and additions to numa.h, which provide
source code compatibility to libnuma version 1.
(The application progamming interface changes, but the ABI is preserved
 through the use of symbol versioning.  So the library stays libnuma.so.1)

numa.h includes numacompat1.h if VERSION1_COMPATIBILITY is defined.

Makefile is changed to install numacompat1.h

The libnuma.c change fixes a bug in numa_node_to_cpus_v2() and adds a
bitmask-to-bitmask copy for unequal-size masks.

The numa.3 man page documents the bitmask copies and the source code
compatibility.

I've tested the source code compatibility by using the version 1
programs in the test suite.

I plan to include this patch (when I get your feedback) and post the
current source as numactl-2.0.0-rc1.tar.gz on
http://oss.sgi.com/projects/libnuma/

Signed-off-by: Cliff Wickman <cpw@sgi.com>