libnuma on big-endian 64-bit systems
Sorry for the long delay. I have now been able to work on this a bit
and come up with a not so broken version of the read_mask function.
The patch you have in 2.0.3-rc1 did not fix the original problem,
but also introduced other bugs, so please revert it.

This is the best I could come up with, haven't tested on little-endian
though because I lack appropriate hardware.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>

 libnuma.c |   40 ++++++++++++++++++++++------------------
 numaint.h |    1 +
 2 files changed, 23 insertions(+), 18 deletions(-)