fix output of numactl --hardware with >BITS_PER_LONG cpus
There was evidently a glitch in the conversion of numactl to libnuma2:

The print_node_cpus function tries to find the size of the bitmask,
which it no longer needs to. This results in printing only the
first word of the CPU mask.

This fixes the function to print all available CPUs again.

Found by Bill Buros on a 128-way 64 bit system.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>