numactl/libnuma - fix parsing of cpu, node mask
[PATCH 05/08] - Fix Parsing of maxproc{cpu|node} in set_thread_constraints()

Against:  numactl-2.0.3-rc2

NOTE:  a patch to fix this has already been posted to linux-numa by
Brice Goglin.  This is the version that I had in my tree.

Hard-coded '+15' was dropping the first character of the mask.
This caused, e.g., numactl to error out on last 4 processors of a
platform--on a 32-cpu system, in my case.

That's a tab separating the tag from the mask, so look for tab
explicitly.

 libnuma.c |   10 ++++++++--
 1 file changed, 8 insertions(+), 2 deletions(-)