libnuma: Fix issue with numactl --hardware on sparse cpumaps
On a machine with sparse cpu ids, numactl fails to print the right cpus:

# numactl --hardware | grep cpus
node 0 cpus: 0 2 4 6
node 1 cpus:

It turns out we were iterating through 0..nr_cpus, not 0..max_cpuid. The
following patch fixes it by using a cpumask instead of an open coded bitmask,
and looping from 0 to the size of the cpumask.

# numactl --hardware | grep cpus
node 0 cpus: 0 2 4 6
node 1 cpus: 8 10 12 14

Signed-off-by: Anton Blanchard <anton@samba.org>