fix cpumask ncpus parameter handling
In Neil's patch:
> +       int ncpumask = (1<<(sysconf(_SC_NPROCESSORS_CONF)))-1;
>         /*
>          * The maximum line size consists of the string at the beginning plus
>          * a digit for each 4 cpus and a comma for each 64 cpus.
> @@ -480,10 +481,22 @@ set_thread_constraints(void)
>         fclose(f);
>         free (buffer);
>
> -       if (maxprocnode < 0) {
> +       /*
> +        * Cps_allowed in the kenrel can be defined to all f's
> +        * i.e. it may be a superset of the actual available processors
> +        * as such lets reduce maxproccpu with a mask of the actual
> +        * available cpus
> +        */
> +       maxproccpu &= ncpumask;

I agree with the fix as long as the number of cpus does not exceed
the number of bits in an int. But that's a pretty small constraint.
I propose to trim maxprocnode to the number of configured cpus - 1.
(And fix the numa_all_cpus_ptr map of all cpus as well)

Then, in testing, I found that read_mask() was buggy.  (the regression
tests don't uncover that).  So that needed some tweaks as well.

Can you give that a try?

Signed-off-by: Cliff Wickman <cpw@sgi.com>
Acked-by: Neil Horman <nhorman@tuxdriver.com>