libnuma/numactl 2.0.4 release announced
On Mon, Jul 26, 2010 at 05:29:21PM -0500, Cliff Wickman wrote:
> libnuma/numactl release 2.0.4 was announced today on freshmeat.net.

I actually found a bug now (looks like exactly one day too late,
how it goes).

The new code does not correct for the -1 quirk in the kernel API for node
masks, unlike the v1 code did. So the last bit in the node masks is going to
be missed.

-Andi