Add IO affinity support to libnuma
This adds new higher level node specifiers to libnuma/numactl.
You can specify IO devices and libnuma automatically resolves
the underlying device if the kernel knows about it.

For example:

netdev:eth0		node of net device eth0
file:/foo/bar 		node of block device /foo/bar is on
ip:hostname		node of net device the route to hostname is pointing to
block:sda3		node of block device sda3 (as known by the kernel)
pci:0.0.1		node of pci device 0.0.1

There are obvious limits to this: in some cases there's no unique node
(like device on RAID). Various IO layers do not export the necessary
information to map back to a PCI device And then on many common systems
the kernel doesn't actually know the node for a PCI device because of
ACPI limitations (but this will hopefully improve).

Still with these caveats it's a useful extension.