libnuma: Fix calculation of maxconfiguredcpu
The value for maxconfiguredcpu was calculated in set_configured_cpus()
by walking the directory in /sys/devices/system/cpu and counting the
entries that begin with "cpu". However, in RHEL6 this ends up making
the cpu count off by two, because there are two directories "cpufreq"
and "cpuidle" that also get counted.

Instead of this overly complex and brittle method, just use sysconf(3)
with _SC_NPROCESSORS_CONF to get the maxconfiguredcpu value. This
was already the fallback method, and seems much less prone to future
failures than other possible methods.

Signed-off-by: Mark A. Grondona <mgrondona@llnl.gov>
Signed-off-by: Petr Holasek <pholasek@redhat.com>