Rewrite numastat to also show per-node process stats
This patch rewrites Andi's numastat perl script as a C program to add features
to show per-node statistics for processes as well as the system as a whole. By
default, the new version is strictly compatible with the old version.  The
behavior changes if any command line options or program arguments are used, as
described in the new man page.  This provides a convenient way to monitor the
NUMA node distribution of memory for various sets of processes. For example, to
check on KVM guests, one might run it like this:

# ./numastat -c qemu

Per-node process memory usage (in MBs)
PID              Node 0 Node 1 Node 2 Node 3 Node 4 Node 5 Node 6 Node 7
Total
---------------  ------ ------ ------ ------ ------ ------ ------ ------
-----
7054 (qemu-kvm)       2      1      0   1391      0      0      0      0
  1395
7174 (qemu-kvm)       2      1      0      1      0      0   1404      0
  1409
7294 (qemu-kvm)       2      1   1389      1      0      0      0      0
  1394
7406 (qemu-kvm)       2      1      0      1      0   1391      0      0
  1395
7530 (qemu-kvm)       2   1401      0      1      0      0      0      0
  1405
7638 (qemu-kvm)       2      1      0      1   1404      0      0      0
  1409
7742 (qemu-kvm)       2      1      0      1      0      0      0   1392
  1397
7858 (qemu-kvm)    1396      1      0      1      0      0      0      0
  1398
---------------  ------ ------ ------ ------ ------ ------ ------ ------
-----
Total              1413   1407   1391   1395   1407   1392   1405   1392
11202

- numastat.8: is rewritten
- numastat: is replaced by numastat.c
- Makefile: is modified to compile numastat.c and to clean numastat

Signed-off-by: Bill Gray <bgray@redhat.com>

 Makefile   |    6
 numastat   |   91 ---
 numastat.8 |  216 +++++---
 numastat.c | 1548 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 4 files changed, 1700 insertions(+), 161 deletions(-)