libnuma: numa_num_task_cpus()
The numa_num_task_cpus() function should return the number of cpu's
available to the caller.  It was returning his highest available cpu
number (+1).
That was an erroneous assumption when the caller is in a cpuset where
available cpu's are not numbered 0..N.

Likewise, numa_num_task_nodes() was doing the same thing.

This patch adds a numa_bitmask_weight() function to count set bits in the map
and makes it available for user convenience as well.

Signed-off-by: Cliff Wickman <cpw@sgi.com>