libnuma: numa_get_run_node_mask and numa_run_on_node_mask
Replace my own erroneous patch: 1010_cw_numa_get_run_node_mask.

That patch replaced numa_get_run_node_mask_v2()'s logic with simply
returning a copy of the bitmask numa_all_nodes_ptr, as that mask is
cpuset aware.
But that mask is not updated when task cpu affinity is changed.
When numactl --show called this function it was getting all the nodes in the
cpuset rather than the affined cpus.
That error was the cause of the regression exposed by test/checkaffinity.
So the original logic of that function (before my patch) is simply enhanced
to only consider nodes within the current cpuset.

And similarly, numa_run_on_node_mask_v2() is enhanced.
(It is given a mask of nodes, and is to create a mask of cpu's that are on
 those nodes.  Then sets affinity to those cpu's.)
It is enhanced to only consider nodes within the current cpuset.

Tested on 96p, 10 nodes. Both inside and outside of a restrictive cpuset.
Tested with 'make test' and the qthreads 'make check'.
The only failure happens when the test/checkaffinity script tries to
set affinity outside of the current cpuset. So that script is fixed in
a separate patch.

Signed-off-by: Cliff Wickman <cpw@sgi.com>