libnuma: test/checkaffinity
Fix test/checkaffinity to be cpuset aware.
This test assumed that both cpus and nodes are numbered 0..N and therefore
can fail simply because it is run in a restrictive cpuset.

Use numactl --show to generate cpuset aware lists of accessible cpus and
nodes.  libnuma uses /proc/self/status to create the lists that --show
displays.

This may seem a bit "chicken and egg", i.e. using libnuma to test libnuma.
We might use /proc/self/cpuset and /dev/cpuset to generate the lists, if
you see a problem with the below approach.

Signed-off-by: Cliff Wickman <cpw@sgi.com>