libnuma: preventing sigbus on touch beyond region
A customer used this example from the numactl.8 man page:
  numactl --offset=1G --length=1G --membind=1 --file /dev/shm/A --touch
intending to cause a 1G allocation of memory on node 1.
But they had only allocated a 1G /dev/shm/A file to begin with.
In this case numactl fails with a SIGBUS.

I don't think this example intended to imply that the --touch would
extend a tmpfs file.  Correct me if I'm wrong.

Either way, numactl should avoid dying from a SIBUS.

This patch:
- Changes the example in numactl.8 for placing a tmpfs file.
- Enhances checks of shm address/offset/length before addressing or mmap'ing
  of the shm region. This prevents invalid mmap calls and SIGBUS's.
- The options to numactl are position dependent, which can confuse a user.
  So the warning messages are clarified.

Signed-off-by: Cliff Wickman <cpw@sgi.com>