libnuma: numa_get_run_node_mask
The numa_get_run_node_mask() function does not do what it is documented
as doing:
  "returns the mask of nodes that the current task is allowed to run on."
It returns a bit mask of cpu's.  And that mask is erroneous.
I've been testing it in conjunction with the qthreads package, which
needs this function to be cpuset-aware.

Does anyone use this function?

The below simply makes a copy of the numa_all_nodes_ptr mask, which really
is a mask of available nodes, and is cpuset-aware.  That is, it is
constrained to the nodes that the caller can use.

Signed-off-by: Cliff Wickman <cpw@sgi.com>

------------- this patch subsequently replaced on 110112 ---------------