test: make test scripts run from any location.
No longer require that the scripts are run from inside the test/
directory.

Introduce two environment variables ${srcdir} and ${builddir} to help
the tests find the other tests, test helper scripts and the built
binaries. Guess the locations of ${srcdir} and ${builddir} in case they
are not defined in the environment.

Tested:
- Made sure "make test" still works.
- Also called the scripts directory from outside the test directory.

Signed-off-by: Filipe Brandenburger <filbranden@google.com>