build: add Travis config for continuous integration
Run configure, build and run some of the tests.

Unfortunately, Travis does not export /sys to continuous integration
jobs, so running the numactl test suite on that platform is not very
useful. Still, it can be used to confirm that the build is not broken,
compiler warnings are not introduced, that the source is compatible with
both gcc and clang and that the distribution archive is cleanly created.

Tested:
- https://travis-ci.org/filbranden/numactl/builds/31826626

Signed-off-by: Filipe Brandenburger <filbranden@google.com>