Clean up numa_error() calls.
It appends a colon and strerror(errno), so avoid newlines at the ends
of messages passed to it.  Also make sure errno is set to something
vaguely sensible before calling it if it's not right after a syscall,
so that we don't just report whatever status happened to be left over
from whatever set it last.