gcc: numa.h: compiler objects to non-const cast of const parameters
/projects/numactl/aarch64/include/numa.h: In function 'nodemask_equal':
/projects/numactl/aarch64/include/numa.h:110:16: error: cast discards 'const' qualifier from pointer target type [-Werror=cast-qual]
  tmp_a.maskp = (unsigned long *)a;

struct bitmask is a general struct that can do const or write ops.

Unfortunately it means when we prepare one with a parameter that
was specifically a const pointer, newer compilers object violently.

Introduce an identical struct const_bitmask which only has a const pointer to
the payload, and cast that to a (struct bitmask *) (or (const struct bitmask *)
if possible) before use.

struct const_bitmask isn't used in any function parameters or outside of
numa.h definitions, so there are no api changes involved.