Revert "make clearcache work on x86/PIC"
This reverts commit 108a9fc0f95f1aca9096334f17430ab7cbab03ed.

The original problem that commit tried to fix is building with -fPIC
on 32-bit x86 architecture caused this compilation error:

  clearcache.c:60:2: error: inconsistent operand constraints in an ‘asm’

That happened because gcc would use the %ebx register on that
architecture to point to the GOT/PLT and wasn't prepared to save/restore
it on its own.

This was a problem on gcc 4, but it has been fixed on gcc 5.

I confirmed that I could reproduce the problem with a small reproducer
on gcc 4.8.2 (CentOS 7). I could no longer reproduce the problem on
gcc 6.4.0 or 7.3.0 (Debian 9).

Some references to the gcc bug and fix commit seem to be here:

  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=54232
  https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=a9d8ab3820ccd10d2ac547295dadf75a33077f14

(Though I'm not 100% sure they're really the ones, they were closest to
this issue.)

The problem with the fix is that it only saves/restores the lower 32-bit
part of the register, so on x86_64 architecture this actually *broke*
numademo, since the `mem` pointer was apparently kept in %ebx and after
the call to clearcache() the upper 32-bits were gone, producing a wild
pointer and crashing with a segmentation fault.

A possible solution would be to guard the xchg solution within an #if
checking for __i386__ and __PIC__ defines, but I'd say there's little
point to that given modern compilers are just able to handle this code
just fine.

Tested: `make check` now passes again on Debian 9 on x86_64.