Allow linking with lld by deduplicating symbols
lld appears to be more strict than GNU ld about duplicate exported
symbols: it appears that inherited symbols in the export specification
from a previous version should not be repeated.

This change removes from version 1.2 of the set of exported symbols
those symbols which are already included in the set for version 1.1.
This allows numactl to be compiled with lld without any errors.

Fixes #46