Drop custom CFLAGS for libnuma itself
Since the LTO build has been fixed with the SYMVER commit (92d4e7d66c7255),
we can now drop the custom CFLAGS for building libnuma sources.

Tested that enabling LTO results in a working build. Tested with:

  $ ./configure \
        CFLAGS='-O2 -g -pipe -Wall -Werror=format-security -Wp,-D_FORTIFY_SOURCE=2 -Wp,-D_GLIBCXX_ASSERTIONS -fexceptions -fstack-protector-strong -grecord-gcc-switches -specs=/usr/lib/rpm/redhat/redhat-hardened-cc1 -specs=/usr/lib/rpm/redhat/redhat-annobin-cc1 -m64 -mtune=generic -fasynchronous-unwind-tables -fstack-clash-protection -fcf-protection -flto' \
        LDFLAGS='-Wl,-z,relro -Wl,--as-needed  -Wl,-z,now -specs=/usr/lib/rpm/redhat/redhat-hardened-ld -flto'

(Note the extra `-flto` at the end of both CFLAGS and LDFLAGS.)

Also compared the symbols in `objdump -T .libs/libnuma.so.1.0.0` of an LTO
build versus a non-LTO build, to confirm that all the expected symbols are
still exported under the expected names.