Use more MPOL_LOCAL
Originally, the way to prefer the local node was to specify
MPOL_PREFERRED with no mask (prior to that MPOL_DEFAULT had that
implicit behavior). However, MPOL_LOCAL interface was added in v3.10
which does the same thing.

	commit 479e2802d09f1e18a97262c4c6f8f17ae5884bd8
	Author: Peter Zijlstra <a.p.zijlstra@chello.nl>
	Date:   Thu Oct 25 14:16:28 2012 +0200

	mm: mempolicy: Make MPOL_LOCAL a real policy

This patch helps readability, local policy use the MPOL_LOCAL.

Unlike MPOL_PREFERRED_MANY, MPOL_LOCAL has existed for 8 years, so we
don't bother trying to detect if it's there automatically.

Signed-off-by: Ben Widawsky <ben.widawsky@intel.com>
Signed-off-by: Feng Tang <feng.tang@intel.com>