build: Refactor autotools setup to store version in separate file
Store the version number in a `VERSION` file at the top of the tree.

This makes it easy to parse the current version by reading the contents
of the file, and also update it (manually or programatically) when
creating a new release.

Use m4 macros to read the version into the autotools configuration script.

Also update it from 2.0.14 to 2.0.15, since a 2.0.15 version was
released but the version inside the configure script was not updated.
(We will soon release a 2.0.16, but let's include this configuration
change there.)

Tested: `make distcheck` passes, confirmed that the `VERSION` file is
included inside the generated `numactl-2.0.15.tar.gz` tarball (which
should make a further `./autogen.sh` run succeed.)

Signed-off-by: Filipe Brandenburger <filbranden@gmail.com>