libnuma: Add APIs for weighted-interleaved allocations
Currently libnuma only supports per-process weighted interleaving
policy. Introduce libnuma APIs to enforce specific memory mappings to be
distributed in the weighted interleave mode.

To align with the existing support for the traditional interleaving,
introduce three APIs, one that interleaves allocations on NUMA nodes
specified in the nodemask, another that interleaves allocations
on all NUMA nodes, and the other that updates the memory policy of an
existing mapping:
  - numa_alloc_weighted_interleaved_subset(size, bitmask)
  - numa_alloc_weighted_interleaved(size)
  - numa_weighted_interleave_memory(start, size, nodemask)

Document the usage of these APIs.

Signed-off-by: Hyeonggon Yoo <hyeonggon.yoo@sk.com>