Add numa_fail_alloc_on_error
Traditionally the numa_alloc_* functions only reported errors for policing
the memory through numa_error/warn. Add a new numa_fail_alloc_on_error
variable that allows to fail the allocation for a policy failure, so that
the caller can handle it.

Default to 0 to not break any existing behavior.

This adds a new shared library symbol name space.

Signed-off-by: Andi Kleen <ak@linux.intel.com>